Bug Fixes, Consistent Code Structure, Comments and More!
This commit (would not be possible without the others) is the one that
really ups the game of Colicious. It enhances the functionality of the
page by fixing various bugs as well as giving the site just an overall
more well rounded / smooth / stable feel.

+ Fixed Github Corner (now links to Github in a new tab, wherever you
press on it, also preventing the trigger of a new color being
generated);
+ When clicking on the hex code to copy the color, now it does not
change the color in the process, this way the background of the
“copied” is not different to the color selected, and its just a better
feel when using the page;
+ Balanced the loading time to be more realistic;
+ Removed a bug where you would accidentally select the hex code or the
button itself when clicking to find a new color;
+ Added support for the backspace and left arrow key to return the same
functionality as the back arrow;
+ Added support for the right arrow key to generate a new color (same
as space);
+ Added comments for better understandability;
+ Structured the code for it to be more readable as well as more easily
understandable;